+++ tests/Integration/Pagination/PaginateDirectiveDBTest.php
@@ -7,6 +7,7 @@ use Illuminate\Database\Eloquent\Collection as EloquentCollection;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Laravel\Scout\Builder as ScoutBuilder;
 use Mockery;
+use Nuwave\Lighthouse\Pagination\Cursor;
 use Tests\DBTestCase;
 use Tests\TestsScoutEngine;
 use Tests\Utils\Models\Comment;
@@ -586,6 +587,114 @@ GRAPHQL;
                             [
                                 'node' => [
                                     'id' => $user->id,
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ])->assertJsonCount(1, 'data.users.edges');
+        });
+    }
+
+    public function testQueriesConnectionPageOffset(): void
+    {
+        $users = factory(User::class, 3)->create();
+
+        $this->schema = /** @lang GraphQL */ '
+        type User {
+            id: ID!
+        }
+
+        type Query {
+            users: [User!]! @paginate(type: CONNECTION)
+        }
+        ';
+
+        $this->assertQueryCountMatches(2, function () use ($users): void {
+            $this->graphQL(/** @lang GraphQL */ '
+            query ($after: String!) {
+                users(first: 2, after: $after) {
+                    pageInfo {
+                      hasNextPage
+                      hasPreviousPage
+                      startCursor
+                      endCursor
+                      total
+                      count
+                      currentPage
+                      lastPage
+                    }
+                    edges {
+                        node {
+                            id
+                        }
+                    }
+                }
+            }
+            ', [
+                'after' => Cursor::encode(2),
+            ])->assertJson([
+                'data' => [
+                    'users' => [
+                        'pageInfo' => [
+                            'hasNextPage' => false,
+                            'hasPreviousPage' => true,
+                            'startCursor' => 'Mw==',
+                            'endCursor' => 'Mw==',
+                            'total' => 3,
+                            'count' => 1,
+                            'currentPage' => 2,
+                            'lastPage' => 2,
+                        ],
+                        'edges' => [
+                            [
+                                'node' => [
+                                    'id' => $users[2]->id,
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ])->assertJsonCount(1, 'data.users.edges');
+        });
+    }
+
+    public function testQueriesConnectionPageOffsetWithoutPageInfo(): void
+    {
+        $users = factory(User::class, 3)->create();
+
+        $this->schema = /** @lang GraphQL */ '
+        type User {
+            id: ID!
+        }
+
+        type Query {
+            users: [User!]! @paginate(type: CONNECTION)
+        }
+        ';
+
+        $cursor = Cursor::encode(2);
+
+        $this->assertQueryCountMatches(1, function () use ($users): void {
+            $this->graphQL(/** @lang GraphQL */ '
+            query ($after: String!) {
+                users(first: 2, after: $after) {
+                    edges {
+                        node {
+                            id
+                        }
+                    }
+                }
+            }
+            ', [
+                'after' => Cursor::encode(2),
+            ])->assertJson([
+                'data' => [
+                    'users' => [
+                        'edges' => [
+                            [
+                                'node' => [
+                                    'id' => $users[2]->id,
                                 ],
                             ],
                         ],